Add files via upload
A Deci-Binary number is a decimal number where the only digits are 1 and 0. For example, 101, 1001, 10011, and 10111001 are all Deci-Binary while 345, 7689, and 1091 are not. This problem asked to return the minimum number of deci-binary numbers required to create a given number [n]. I came up with a fairly elegant solution that I am quite proud of.